+++ app.js
@@ -4,5 +4,5 @@ import Chat from './chat/main'
 new Chat({
   el: '.chat',
   buttonEl: '.website', // element to append chat toogle button. Must be relative.
-  isOpenedOnStart: true // default value
+  isOpenedOnStart: true // default value: true
 })