+++ chat/services/storeService.js
@@ -1,12 +1,21 @@
-class StoreService {
-  constructor () {
-    this.name = 'StoreService'
-  }
+/* eslint-disable no-undef */
 
-  get (key) {
-  }
-  set (key, value) {
+export const storeService = {
+  getItem: (key) => {
+    return localStorage.getItem(key) || ''
+  },
+  setItem: (key, string) => {
+    localStorage.setItem(key, string)
+  },
+
+  remove: (key) => {
+    localStorage.removeItem(key)
+  },
+
+  getJSON: (key) => {
+    return JSON.parse(localStorage.getItem(key) || '[]')
+  },
+  setJSON: (key, obj) => {
+    localStorage.setItem(key, JSON.stringify(obj))
   }
 }
-
-export default StoreService