+++ CHANGELOG.md
@@ -16,6 +16,7 @@ There is no Semantic Versioning yet
 - [x] chat with me button closed setting (default: false)
 - [x] show/hide button component with option el
 - [x] Optimize toggle chat
+- [x] Botik class
 
 #### 0.0.1 - 2017-03-13